# WebService/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Mini Product Ordering Service API.",
        "title": "Gin Mini Product Ordering Service",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/Transaction/All": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Show  All User Transactions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Show  All User Transactions",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "number",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "number",
                        "description": "data per page",
                        "name": "recordPerPage",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/Transaction/Purchase/{uID}": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Purchase All Product from Cart",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Purchase All Product from Cart",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Purchase All Product from Cart ",
                        "name": "uID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/Transaction/Rating": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Add Product Rating",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Add Product Rating",
                "parameters": [
                    {
                        "description": "Add Product Rating",
                        "name": "rating",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ProductRating"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/Transaction/{tID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Detai Transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Transaction"
                ],
                "summary": "Detai Transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Detai Transaction",
                        "name": "tID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/User/delete/{uid}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Delete Account",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Delete Account",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Delete Account by Email",
                        "name": "uid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/User/image": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Post Form",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Image Form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id",
                        "name": "id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "image",
                        "name": "image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/User/profile": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Show User Form",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Show User Form",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Show User by ID",
                        "name": "uid",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/User/profile/all": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Show User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Show User",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "number",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "number",
                        "description": "data per page",
                        "name": "recordPerPage",
                        "in": "query"
                    }
                ],
                "responses": {}
            }
        },
        "/User/update": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update Account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update Account",
                "parameters": [
                    {
                        "description": "Update User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateModel"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/User/verification": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Account Verification Form",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Account Verification Form",
                "parameters": [
                    {
                        "description": "Account Verification User",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AccountVerificationData"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/User/verification/{email}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Account Verification User",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Account Verification User",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Account Verification User by Email",
                        "name": "email",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.AccountVerificationData": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
...
        "models.ProductRating": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "product_id": {
                    "type": "string"
                },
                "rating": {
                    "type": "integer"
                },
                "transaction_id": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateModel": {
            "type": "object",
            "properties": {
                "about": {
                    "type": "string"
                },
                "birthdate": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Token",
            "in": "header"
        }
    }
}